Integration tests for portfolio and funds
